башенки наносят урон (надо подумать, как обработать несколько событий урона за один кадр)

+++ Assets/Scripts/Components/Towers/LaserTurretComponent.cs
@@ -9,7 +9,7 @@ namespace Components.Towers
         public Transform turret;
         public Transform laser;
 
-        public int damagePerSecond;
+        public float damagePerSecond;
         public int attackRadius;
     }
 }
